Split dev-tools/globals* into sim and server
