+++ tsconfig.json
@@ -16,6 +16,7 @@
         "./server/chat-commands.js"
     ],
     "include": [
+        "*.ts",
         "./config/formats.js",
         "./data/*",
         "./data/mods/*/*",
